backend/pyproject.toml
[project]
name = "ml-starter"
version = "0.1.0"
description = "Structure pro FastAPI + ML avec endpoint de prédiction"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic-settings>=2.6.0",
    "joblib>=1.4.0",
    "numpy>=2.0.0",
    "pandas>=3.0.1",
    "scikit-learn>=1.8.0",
    "seaborn>=0.13.2",
    "matplotlib>=3.10.8",
    "pytest>=9.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.28.0",
    "ruff>=0.11.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
]

[tool.uv]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
addopts = "-q --cov=app --cov-report=term-missing"

[tool.ruff]
target-version = "py312"
line-length = 100
src = [".", "app", "tests", "scripts"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "SIM", "RUF"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"

[tool.mypy]
python_version = "3.12"
pretty = true
strict = true
warn_unused_configs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
show_error_codes = true
exclude = ["build/"]

[[tool.mypy.overrides]]
module = ["sklearn.*", "joblib"]
ignore_missing_imports = true

[tool.coverage.run]
source = ["app"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false

[dependency-groups]
dev = []
